SETRAGESTION/protected_excel/requirements.txt
streamlit
pandas
plotly
reportlab
Pillow
cloudinary
requests
python-dotenv
matplotlib
openpyxl
pystray
psutil